fix: the file 'StocksAvailableData' was renamed to 'StocksAvailableToInvestData' and the data from the 'StocksAvailableToInvestData' file was imported into the 'StocksAvailableToInvestTable' file
