+++ src/utils/StocksAvailableToInvestData.js
